pyproject.toml
[project]
name = "comicdownloader"
version = "0.1.0"
description = "A simple CLI tool for downloading comic books from the web"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.12"
dependencies = [
    "curl-cffi==0.7.3",
    "pillow==10.4.0",
    "selectolax==0.3.24",
]

[project.scripts]
comicdownloader = "comicdownloader:main"
